Improve network manager interface
