+++ finality-aleph/src/party/mod.rs
@@ -154,7 +154,7 @@ where
                 .session_manager
                 .start_nonvalidator_session(session_id, authorities)
             {
-                warn!(target: "aleph-party", "Failed to start nonvalidator session{:?}:{:?}", session_id, e);
+                warn!(target: "aleph-party", "Failed to start nonvalidator session{:?}: {}", session_id, e);
             }
             None
         };
@@ -194,21 +194,22 @@ where
                 } => {
                     let next_session_authorities = next_session_authority_data.authorities();
                     match self.session_manager.node_idx(next_session_authorities).await {
-                         Some(_) => if let Err(e) = self
+                         Some(next_session_node_id) => if let Err(e) = self
                                 .session_manager
                                 .early_start_validator_session(
                                     next_session_id,
+                                    next_session_node_id,
                                     next_session_authorities,
                                 ).await
                             {
-                                warn!(target: "aleph-party", "Failed to early start validator session{:?}:{:?}", next_session_id, e);
+                                warn!(target: "aleph-party", "Failed to early start validator session{:?}: {}", next_session_id, e);
                             }
                         None => {
                             if let Err(e) = self
                                 .session_manager
                                 .start_nonvalidator_session(next_session_id, next_session_authorities)
                             {
-                                warn!(target: "aleph-party", "Failed to early start nonvalidator session{:?}:{:?}", next_session_id, e);
+                                warn!(target: "aleph-party", "Failed to early start nonvalidator session{:?}: {}", next_session_id, e);
                             }
                         }
                     }
@@ -232,7 +233,7 @@ where
             }
         }
         if let Err(e) = self.session_manager.stop_session(session_id) {
-            warn!(target: "aleph-party", "Session Manager failed to stop in session {:?}: {:?}", session_id, e)
+            warn!(target: "aleph-party", "Session Manager failed to stop in session {:?}: {}", session_id, e)
         }
     }
